requirements.txt
django
djangorestframework
psycopg2
docker-compose